fix(taxis): correct oikonomos dep name + fix dead_code + struct docs + type alignment (#5482)
The auto-authored refactor (052f332) added five compilation blockers:

1. Cargo.toml: `daemon = {..}` references the package at crates/daemon/ whose
   Cargo.toml declares `name = "oikonomos"`. Rename the dev-dep key so cargo
   can resolve it.

2. eidos/src/knowledge/mod.rs: `fact` was a private module; taxis tests
   reference `eidos::knowledge::fact::MAX_CONTENT_LENGTH`. Make it `pub mod`.

3. organon/src/builtins/memory/mod.rs: `datalog` was private; taxis tests
   reference `organon::builtins::memory::datalog::DEFAULT_{ROW_LIMIT,
   TIMEOUT_SECS}`. Make it `pub mod` (also fixes dead_code on the pub consts).

4. Behavior file struct-doc placement: kimi inserted a struct-description
   `///` block before the constants instead of before the struct, then left a
   blank line between them. Clippy empty_line_after_doc_comments fired, and a
   missing_docs followed once the blank was collapsed. Reorder: constants come
   first (each with its own `///` doc), struct doc sits directly above
   `#[derive(…)]`.

5. Type mismatches and lint cascade: `DEFAULT_TOOL_DATALOG_DEFAULT_ROW_LIMIT`
   was u32 (should be usize to match organon); `DEFAULT_TOOL_MAX_IMAGE_BYTES`
   and `DEFAULT_TOOL_MAX_PDF_BYTES` were usize (should be u64 to match
   organon::builtins::view_file consts). Aligned types eliminate all as_conversions
   and cast_possible_truncation lints. Poll-interval assertion changed from
   `.as_millis() as u64` to `.as_secs() * 1_000` (avoids u128 cast entirely).
   `mirrored_defaults_match_foreign_constants` split into six focused sub-tests
   to satisfy too_many_lines; float_cmp suppressed with #[expect(…)] where
   compile-time constants are compared by definition-equal values.

All 52 formerly-mirrored fields are now enforced by compile-time assert!() or
direct constant reference. 364/364 taxis nextest pass.

+++ crates/taxis/src/config/agents.rs
@@ -13,13 +13,13 @@ use super::{AgencyLevel, BookkeepingProviderKind, CompactionStrategyKind};
 /// Default value used for `AgentBehaviorDefaults::tool_agent_dispatch_max_tasks`.
 pub(crate) const DEFAULT_TOOL_AGENT_DISPATCH_MAX_TASKS: usize = 10;
 /// Default value used for `AgentBehaviorDefaults::tool_datalog_default_row_limit`.
-pub(crate) const DEFAULT_TOOL_DATALOG_DEFAULT_ROW_LIMIT: u32 = 100;
+pub(crate) const DEFAULT_TOOL_DATALOG_DEFAULT_ROW_LIMIT: usize = 100;
 /// Default value used for `AgentBehaviorDefaults::tool_datalog_default_timeout_secs`.
 pub(crate) const DEFAULT_TOOL_DATALOG_DEFAULT_TIMEOUT_SECS: f64 = 5.0;
 /// Default value used for `AgentBehaviorDefaults::tool_max_image_bytes`.
-pub(crate) const DEFAULT_TOOL_MAX_IMAGE_BYTES: usize = 20_971_520;
+pub(crate) const DEFAULT_TOOL_MAX_IMAGE_BYTES: u64 = 20_971_520;
 /// Default value used for `AgentBehaviorDefaults::tool_max_pdf_bytes`.
-pub(crate) const DEFAULT_TOOL_MAX_PDF_BYTES: usize = 33_554_432;
+pub(crate) const DEFAULT_TOOL_MAX_PDF_BYTES: u64 = 33_554_432;
 
 /// Agent configuration: shared defaults and per-agent definitions.
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -649,13 +649,13 @@ pub struct AgentBehaviorDefaults { // kanon:ignore RUST/struct-too-many-fields â
     /// Maximum concurrent agent-dispatch tasks.
     pub tool_agent_dispatch_max_tasks: usize,
     /// Default row limit for Datalog memory queries.
-    pub tool_datalog_default_row_limit: u32,
+    pub tool_datalog_default_row_limit: usize,
     /// Default query timeout in seconds for the Datalog memory tool.
     pub tool_datalog_default_timeout_secs: f64,
     /// Maximum image file size in bytes for the view-file tool.
-    pub tool_max_image_bytes: usize,
+    pub tool_max_image_bytes: u64,
     /// Maximum PDF file size in bytes for the view-file tool.
-    pub tool_max_pdf_bytes: usize,
+    pub tool_max_pdf_bytes: u64,
 
     // --- Bootstrap ---
     /// Minimum token budget remaining before attempting section truncation.
@@ -780,8 +780,7 @@ const _: () =
     assert!(DEFAULT_TOOL_AGENT_DISPATCH_MAX_TASKS == organon::builtins::agent::MAX_DISPATCH_TASKS);
 #[cfg(test)]
 const _: () = assert!(
-    DEFAULT_TOOL_DATALOG_DEFAULT_ROW_LIMIT as usize
-        == organon::builtins::memory::datalog::DEFAULT_ROW_LIMIT
+    DEFAULT_TOOL_DATALOG_DEFAULT_ROW_LIMIT == organon::builtins::memory::datalog::DEFAULT_ROW_LIMIT
 );
 #[cfg(test)]
 const _: () = assert!(
@@ -790,7 +789,6 @@ const _: () = assert!(
 );
 #[cfg(test)]
 const _: () =
-    assert!(DEFAULT_TOOL_MAX_IMAGE_BYTES as u64 == organon::builtins::view_file::MAX_IMAGE_BYTES);
+    assert!(DEFAULT_TOOL_MAX_IMAGE_BYTES == organon::builtins::view_file::MAX_IMAGE_BYTES);
 #[cfg(test)]
-const _: () =
-    assert!(DEFAULT_TOOL_MAX_PDF_BYTES as u64 == organon::builtins::view_file::MAX_PDF_BYTES);
+const _: () = assert!(DEFAULT_TOOL_MAX_PDF_BYTES == organon::builtins::view_file::MAX_PDF_BYTES);
